SnapCamera: show previewcover when surfaceview size changed
issue: Half black screen is seen for few seconds when recording is
started
fix: When start recording or stop recording , the SurfaceView size is changed,
the content on it is not cleared. So we show the previewcover to cover the content.
And hide the cover when new preview frame arrived.

Change-Id: I485880335de12008e39d92f5093d06c56a8df759